python-adk-agent/requirements.txt
google-adk
requests
sseclient-py